+++ modules/gui/src/cpp/QtMainWindow.h
@@ -71,7 +71,7 @@ private slots:
     changeDir();
 
     void
-    editor();
+    textEditor();
 
     void
     historyBrowserToggle();