+++ modules/text_editor/src/cpp/TextEditor.cpp
@@ -18,7 +18,7 @@ namespace Nelson {
 static QtTextEditor* edit = nullptr;
 //=============================================================================
 bool
-editor(Evaluator* eval)
+textEditor(Evaluator* eval)
 {
     bool res = false;
     if (edit == nullptr) {
@@ -34,7 +34,7 @@ editor(Evaluator* eval)
 }
 //=============================================================================
 bool
-editor(Evaluator* eval, bool importSharedText)
+textEditor(Evaluator* eval, bool importSharedText)
 {
     if (edit == nullptr) {
         edit = new QtTextEditor(eval);
@@ -54,7 +54,7 @@ editor(Evaluator* eval, bool importSharedText)
 }
 //=============================================================================
 bool
-editor(Evaluator* eval, const std::wstring& filename)
+textEditor(Evaluator* eval, const std::wstring& filename)
 {
     bool res = false;
     if (edit == nullptr) {
@@ -71,17 +71,17 @@ editor(Evaluator* eval, const std::wstring& filename)
 }
 //=============================================================================
 bool
-editor(Evaluator* eval, const wstringVector& filenames)
+textEditor(Evaluator* eval, const wstringVector& filenames)
 {
     bool res = true;
     for (const auto& filename : filenames) {
-        res = res && editor(eval, filename);
+        res = res && textEditor(eval, filename);
     }
     return res;
 }
 //=============================================================================
 bool
-closeEditor()
+closeTextEditor()
 {
     if (edit != nullptr) {
         delete edit;
@@ -89,6 +89,25 @@ closeEditor()
         return true;
     }
     return false;
+}
+//=============================================================================
+bool
+isTextEditorVisible()
+{
+    if (edit != nullptr) {
+        return edit->isVisible();
+    }
+    return false;
+}
+//=============================================================================
+void
+showTextEditor()
+{
+    if (edit != nullptr) {
+        edit->show();
+        edit->activateWindow();
+        edit->setWindowState(Qt::WindowActive);
+    }
 }
 //=============================================================================
 }